# [Performance] reduce calls to exponential_offset_ge